# docs/apis/account-and-transaction/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Account and Transaction API",
    "description": "Account and Transaction API for Open Banking Connector Web App",
    "version": "web-app/5.1.0"
  },
  "paths": {
...
    "/aisp/standing-orders": {
      "get": {
        "tags": [
          "Standing Orders"
        ],
        "summary": "Read standing orders",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StandingOrdersResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandingOrdersResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandingOrdersResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/accounts/{externalApiAccountId}/standing-orders": {
      "get": {
        "tags": [
          "Standing Orders"
        ],
        "summary": "Read standing orders",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StandingOrdersResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandingOrdersResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandingOrdersResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Read transactions",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalApiStatementId",
            "in": "path",
            "description": "External (bank) API ID of Statement",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromBookingDateTime",
            "in": "query",
            "description": "Start date and time for filtering of the Transaction records on the\r\nTransaction/BookingDateTime field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toBookingDateTime",
            "in": "query",
            "description": "End date and time for filtering of the Transaction records on the\r\nTransaction/BookingDateTime field",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/accounts/{externalApiAccountId}/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Read transactions",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalApiStatementId",
            "in": "path",
            "description": "External (bank) API ID of Statement",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromBookingDateTime",
            "in": "query",
            "description": "Start date and time for filtering of the Transaction records on the\r\nTransaction/BookingDateTime field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toBookingDateTime",
            "in": "query",
            "description": "End date and time for filtering of the Transaction records on the\r\nTransaction/BookingDateTime field",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/accounts/{externalApiAccountId}/statements/{externalApiStatementId}/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Read transactions",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalApiStatementId",
            "in": "path",
            "description": "External (bank) API ID of Statement",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromBookingDateTime",
            "in": "query",
            "description": "Start date and time for filtering of the Transaction records on the\r\nTransaction/BookingDateTime field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toBookingDateTime",
            "in": "query",
            "description": "End date and time for filtering of the Transaction records on the\r\nTransaction/BookingDateTime field",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Links": {
        "type": "object",
        "properties": {
          "Self": {
            "type": "string",
            "description": "Gets the self.",
            "nullable": true
          },
          "First": {
            "type": "string",
            "description": "Gets the first.",
            "nullable": true
          },
          "Prev": {
            "type": "string",
            "description": "Gets the prev.",
            "nullable": true
          },
          "Next": {
            "type": "string",
            "description": "Gets the next.",
            "nullable": true
          },
          "Last": {
            "type": "string",
            "description": "Gets the last.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Links relevant to the payload."
      },
      "Meta": {
        "type": "object",
        "properties": {
          "TotalPages": {
            "type": "integer",
            "description": "Gets the total pages.",
            "format": "int32",
            "nullable": true
          },
          "FirstAvailableDateTime": {
            "type": "string",
            "description": "All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          },
          "LastAvailableDateTime": {
            "type": "string",
            "description": "All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Meta Data relevant to the payload."
      },
...
      "OBActiveOrHistoricCurrencyAndAmount2": {
        "type": "object",
        "properties": {
          "Amount": {
            "type": "string",
            "description": "A number of monetary units specified in an active currency where the unit of currency is explicit and compliant with ISO 4217.",
            "nullable": true
          },
          "Currency": {
            "type": "string",
            "description": "A code allocated to a currency by a Maintenance Agency under an international identification scheme, as described in the latest edition of the international standard ISO 4217 \"Codes for the representation of currencies and funds\".",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The amount of the first Standing Order."
      },
...
      "OBCreditDebitCode1Enum": {
        "enum": [
          "Credit",
          "Debit"
        ],
        "type": "string",
        "description": "Indicates whether the transaction is a credit or a debit entry."
      },
...
      "OBEntryStatus1CodeEnum": {
        "enum": [
          "Booked",
          "Pending",
          "Rejected"
        ],
        "type": "string",
        "description": "Status of a transaction entry on the books of the account servicer."
      },
...
      "OBExternalStandingOrderStatus1CodeEnum": {
        "enum": [
          "Active",
          "Inactive"
        ],
        "type": "string",
        "description": "Specifies the status of the standing order in code form."
      },
...
      "OBReadDataTransaction6": {
        "type": "object",
        "properties": {
          "Transaction": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OBTransaction6"
            },
            "description": "Gets the transaction.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The OBReadDataTransaction6."
      },
...
      "OBReadStandingOrder6": {
        "type": "object",
        "properties": {
          "Data": {
            "$ref": "#/components/schemas/OBReadStandingOrder6Data"
          },
          "Links": {
            "$ref": "#/components/schemas/Links"
          },
          "Meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "description": "The OBReadStandingOrder6."
      },
      "OBReadStandingOrder6Data": {
        "type": "object",
        "properties": {
          "StandingOrder": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OBStandingOrder6"
            },
            "description": "Gets the standing order.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The OBReadStandingOrder6Data."
      },
      "OBReadTransaction6": {
        "type": "object",
        "properties": {
          "Data": {
            "$ref": "#/components/schemas/OBReadDataTransaction6"
          },
          "Links": {
            "$ref": "#/components/schemas/Links"
          },
          "Meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "description": "Response object OBReadTransaction6 from UK Open Banking Read-Write Account and Transaction API spec\r\n<a href=\"https://github.com/OpenBankingUK/read-write-api-specs/blob/v3.1.10r5/dist/openapi/account-info-openapi.yaml\" />\r\nv3.1.10r5 <a />. Open Banking Connector will automatically\r\ntranslate <i>to</i> this from an older format for banks supporting an earlier spec version."
      },
      "OBStandingOrder6": {
        "type": "object",
        "properties": {
          "AccountId": {
            "type": "string",
            "description": "A unique and immutable identifier used to identify the account resource. This identifier has no meaning to the account owner.",
            "nullable": true
          },
          "StandingOrderId": {
            "type": "string",
            "description": "A unique and immutable identifier used to identify the standing order resource. This identifier has no meaning to the account owner.",
            "nullable": true
          },
          "Frequency": {
            "type": "string",
            "description": "Individual Definitions:\r\nNotKnown - Not Known\r\nEvryDay - Every day\r\nEvryWorkgDay - Every working day\r\nIntrvlDay - An interval specified in number of calendar days (02 to 31)\r\nIntrvlWkDay - An interval specified in weeks (01 to 09), and the day within the week (01 to 07)\r\nWkInMnthDay - A monthly interval, specifying the week of the month (01 to 05) and day within the week (01 to 07)\r\nIntrvlMnthDay - An interval specified in months (between 01 to 06, 12, 24), specifying the day within the month (-05 to -01, 01 to 31)\r\nQtrDay - Quarterly (either ENGLISH, SCOTTISH, or RECEIVED)\r\nENGLISH = Paid on the 25th March, 24th June, 29th September and 25th December.\r\nSCOTTISH = Paid on the 2nd February, 15th May, 1st August and 11th November.\r\nRECEIVED = Paid on the 20th March, 19th June, 24th September and 20th December.\r\nIndividual Patterns:\r\nNotKnown (ScheduleCode)\r\nEvryDay (ScheduleCode)\r\nEvryWorkgDay (ScheduleCode)\r\nIntrvlDay:NoOfDay (ScheduleCode + NoOfDay)\r\nIntrvlWkDay:IntervalInWeeks:DayInWeek (ScheduleCode + IntervalInWeeks + DayInWeek)\r\nWkInMnthDay:WeekInMonth:DayInWeek (ScheduleCode + WeekInMonth + DayInWeek)\r\nIntrvlMnthDay:IntervalInMonths:DayInMonth (ScheduleCode + IntervalInMonths + DayInMonth)\r\nQtrDay: + either (ENGLISH, SCOTTISH or RECEIVED) ScheduleCode + QuarterDay\r\nThe regular expression for this element combines five smaller versions for each permitted pattern. To aid legibility - the components are presented individually here:\r\nNotKnown\r\nEvryDay\r\nEvryWorkgDay\r\nIntrvlDay:((0[2-9])|([1-2][0-9])|3[0-1])\r\nIntrvlWkDay:0[1-9]:0[1-7]\r\nWkInMnthDay:0[1-5]:0[1-7]\r\nIntrvlMnthDay:(0[1-6]|12|24):(-0[1-5]|0[1-9]|[12][0-9]|3[01])\r\nQtrDay:(ENGLISH|SCOTTISH|RECEIVED)\r\nFull Regular Expression:\r\n^(NotKnown)$|^(EvryDay)$|^(EvryWorkgDay)$|^(IntrvlDay:((0[2-9])|([1-2][0-9])|3[0-1]))$|^(IntrvlWkDay:0[1-9]:0[1-7])$|^(WkInMnthDay:0[1-5]:0[1-7])$|^(IntrvlMnthDay:(0[1-6]|12|24):(-0[1-5]|0[1-9]|[12][0-9]|3[01]))$|^(QtrDay:(ENGLISH|SCOTTISH|RECEIVED))$",
            "nullable": true
          },
          "Reference": {
            "type": "string",
            "description": "Unique reference, as assigned by the creditor, to unambiguously refer to the payment transaction.\r\nUsage: If available, the initiating party should provide this reference in the structured remittance information, to enable reconciliation by the creditor upon receipt of the amount of money.\r\nIf the business context requires the use of a creditor reference or a payment remit identification, and only one identifier can be passed through the end-to-end chain, the creditor's reference or payment remittance identification should be quoted in the end-to-end transaction identification.",
            "nullable": true
          },
          "FirstPaymentDateTime": {
            "type": "string",
            "description": "The date on which the first payment for a Standing Order schedule will be made.All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          },
          "NextPaymentDateTime": {
            "type": "string",
            "description": "The date on which the next payment for a Standing Order schedule will be made.All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          },
          "LastPaymentDateTime": {
            "type": "string",
            "description": "The date on which the last (most recent) payment for a Standing Order schedule was made.All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          },
          "FinalPaymentDateTime": {
            "type": "string",
            "description": "The date on which the final payment for a Standing Order schedule will be made.All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          },
          "NumberOfPayments": {
            "type": "string",
            "description": "Number of the payments that will be made in completing this frequency sequence including any executed since the sequence start date.",
            "nullable": true
          },
          "StandingOrderStatusCode": {
            "$ref": "#/components/schemas/OBExternalStandingOrderStatus1CodeEnum"
          },
          "FirstPaymentAmount": {
            "$ref": "#/components/schemas/OBActiveOrHistoricCurrencyAndAmount2"
          },
          "NextPaymentAmount": {
            "$ref": "#/components/schemas/OBActiveOrHistoricCurrencyAndAmount3"
          },
          "LastPaymentAmount": {
            "$ref": "#/components/schemas/OBActiveOrHistoricCurrencyAndAmount11"
          },
          "FinalPaymentAmount": {
            "$ref": "#/components/schemas/OBActiveOrHistoricCurrencyAndAmount4"
          },
          "CreditorAgent": {
            "$ref": "#/components/schemas/OBBranchAndFinancialInstitutionIdentification51"
          },
          "CreditorAccount": {
            "$ref": "#/components/schemas/OBCashAccount51"
          },
          "SupplementaryData": {
            "type": "object",
            "additionalProperties": { },
            "description": "Additional information that can not be captured in the structured fields and/or any other specific block.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The OBStandingOrder6."
      },
      "OBTransaction6": {
        "type": "object",
        "properties": {
          "AccountId": {
            "type": "string",
            "description": "A unique and immutable identifier used to identify the account resource. This identifier has no meaning to the account owner.",
            "nullable": true
          },
          "TransactionId": {
            "type": "string",
            "description": "Unique identifier for the transaction within an servicing institution. This identifier is both unique and immutable.",
            "nullable": true
          },
          "TransactionReference": {
            "type": "string",
            "description": "Unique reference for the transaction. This reference is optionally populated, and may as an example be the FPID in the Faster Payments context.",
            "nullable": true
          },
          "StatementReference": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets the statement reference.",
            "nullable": true
          },
          "CreditDebitIndicator": {
            "$ref": "#/components/schemas/OBCreditDebitCode1Enum"
          },
          "Status": {
            "$ref": "#/components/schemas/OBEntryStatus1CodeEnum"
          },
          "TransactionMutability": {
            "$ref": "#/components/schemas/OBTransactionMutability1CodeEnum"
          },
          "BookingDateTime": {
            "type": "string",
            "description": "Date and time when a transaction entry is posted to an account on the account servicer's books.\r\nUsage: Booking date is the expected booking date, unless the status is booked, in which case it is the actual booking date.All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time"
          },
          "ValueDateTime": {
            "type": "string",
            "description": "Date and time at which assets become available to the account owner in case of a credit entry, or cease to be available to the account owner in case of a debit transaction entry.\r\nUsage: If transaction entry status is pending and value date is present, then the value date refers to an expected/requested value date.\r\nFor transaction entries subject to availability/float and for which availability information is provided, the value date must not be used. In this case the availability component identifies the number of availability days.All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          },
          "TransactionInformation": {
            "type": "string",
            "description": "Further details of the transaction. \r\nThis is the transaction narrative, which is unstructured text.",
            "nullable": true
          },
          "AddressLine": {
            "type": "string",
            "description": "Information that locates and identifies a specific address for a transaction entry, that is presented in free format text.",
            "nullable": true
          },
          "Amount": {
            "$ref": "#/components/schemas/OBActiveOrHistoricCurrencyAndAmount9"
          },
          "ChargeAmount": {
            "$ref": "#/components/schemas/OBActiveOrHistoricCurrencyAndAmount10"
          },
          "CurrencyExchange": {
            "$ref": "#/components/schemas/OBCurrencyExchange5"
          },
          "BankTransactionCode": {
            "$ref": "#/components/schemas/OBBankTransactionCodeStructure1"
          },
          "ProprietaryBankTransactionCode": {
            "$ref": "#/components/schemas/ProprietaryBankTransactionCodeStructure1"
          },
          "Balance": {
            "$ref": "#/components/schemas/OBTransactionCashBalance"
          },
          "MerchantDetails": {
            "$ref": "#/components/schemas/OBMerchantDetails1"
          },
          "CreditorAgent": {
            "$ref": "#/components/schemas/OBBranchAndFinancialInstitutionIdentification61"
          },
          "CreditorAccount": {
            "$ref": "#/components/schemas/OBCashAccount60"
          },
          "DebtorAgent": {
            "$ref": "#/components/schemas/OBBranchAndFinancialInstitutionIdentification62"
          },
          "DebtorAccount": {
            "$ref": "#/components/schemas/OBCashAccount61"
          },
          "CardInstrument": {
            "$ref": "#/components/schemas/OBTransactionCardInstrument1"
          },
          "SupplementaryData": {
            "type": "object",
            "additionalProperties": { },
            "description": "Additional information that can not be captured in the structured fields and/or any other specific block.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Provides further details on an entry in the report."
      },
...
      "StandingOrdersResponse": {
        "type": "object",
        "properties": {
          "ExternalApiResponse": {
            "$ref": "#/components/schemas/OBReadStandingOrder6"
          },
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to Standing Order read requests"
      },
      "TransactionsResponse": {
        "type": "object",
        "properties": {
          "ExternalApiResponse": {
            "$ref": "#/components/schemas/OBReadTransaction6"
          },
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to Transaction read requests"
      }
    }
  }
}